Filter custom datasets to only read .json files
By default, the custom dataset preperation code read all files in the
specified directory as opposed to just .json, .jsonl, or .json.gz files.
This could cause errors if attempting to preprocess a directory that has
already been run once (even with failures) as it creates extra
files/directories as well as directories that contain non-json files in
it.

Signed-Off-By: Robert Clark <roclark@nvidia.com>