Remove bcp_launcher for data preprocessing
Explicitly calling mpirun as the launcher on BCP clusters for data
preparation can significantly slow down the processing time as mpirun
on BCP locks each process to a single core instead of using all
available cores on the system. Allowing BCP to manage the processes
itself can yield as much as 30x higher throughput for the same
functionality.

Signed-Off-By: Robert Clark <roclark@nvidia.com>