Allow imagePullSecrets to be optional
Certain cluster schedulers, such as Run:ai, handle imagePullSecrets
independently of what's listed in the manifest file. Additonally, it's
possible that users could push a custom NeMo image to another container
registry that doesn't require authentication. As such, the ability to
omit the imagePullSecret from the config should be supported to allow
users to run jobs on kubernetes without it.

Signed-Off-By: Robert Clark <roclark@nvidia.com>