Remove dup line in `llama3_1_8b.yaml`
This line is a duplicate of the previous line. If it is not removed, it will cause an omegaconf error with the message 'found duplicate key sequence_parallel'. Alternatively, could you explain why it is necessary to have an independent 'llama3_1_xb' configuration when the 'llama3_xb' configuration files already exist?