Set k8s imagePullPolicy to Always
To prevent certain corner cases where images could be pushed with the
same tag to a container registry, the imagePullPolicy for k8s jobs
should use "Always" to ensure the container registry is always checked
for newer versions of the container, even if one is already cached
locally. This shouldn't negatively impact performance as it won't
re-pull an image if it already exists locally and there isn't a newer
one available on the registry.

Signed-Off-By: Robert Clark <roclark@nvidia.com>